requirements.txt
keyring
keyrings.alt
numpy
requests
Pillow
slack-bolt
transformers